Deprecate compat.is_<platform>, rewriting all uses

+++ git/objects/submodule/base.py
@@ -10,7 +10,7 @@ import uuid
 
 import git
 from git.cmd import Git
-from git.compat import defenc, is_win
+from git.compat import defenc
 from git.config import GitConfigParser, SectionConstraint, cp
 from git.exc import (
     BadName,
@@ -356,9 +356,8 @@ class Submodule(IndexObject, TraversableIterableObj):
         """
         git_file = osp.join(working_tree_dir, ".git")
         rela_path = osp.relpath(module_abspath, start=working_tree_dir)
-        if is_win:
-            if osp.isfile(git_file):
-                os.remove(git_file)
+        if os.name == "nt" and osp.isfile(git_file):
+            os.remove(git_file)
         with open(git_file, "wb") as fp:
             fp.write(("gitdir: %s" % rela_path).encode(defenc))
